fix(roles): enforce tx authority
